Move `JavadocEscaper` to `javadoc` package
